+++ core/src/main/java/com/graphhopper/routing/ev/DefaultEncodedValueFactory.java
@@ -97,6 +97,8 @@ public class DefaultEncodedValueFactory implements EncodedValueFactory {
             return new EnumEncodedValue<>(Crossing.KEY, Crossing.class);
         } else if (CarConditionalAccess.KEY.equals(name)) {
             return CarConditionalAccess.create();
+        } else if (FerrySpeed.KEY.equals(name)) {
+            return FerrySpeed.create();
         } else {
             throw new IllegalArgumentException("DefaultEncodedValueFactory cannot find EncodedValue " + name);
         }