+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 8.0 [not yet released]
 
 - conditional access restriction tags are no longer considered from vehicle tag parsers and instead a new car_conditional_access encoded value can be used in a custom model. Seasonal restrictions will be ignored to handle problems described in #2477 but could be shown with path details "access_conditional". See #1234
+- removed duration:seconds as intermediate tag
 - /info endpoint does no longer return the vehicle used per profile and won't return encoded value of vehicles like car_average_speed
 - Country rules no longer contain maxspeed handling, enable a much better alternative via `max_speed_calculator.enabled: true`. On the client side use `max_speed_estimated` to determine if max_speed is from OSM or an estimation. See #2810
 - bike routing better avoids dangerous roads, see #2796 and #2802
@@ -14,6 +15,8 @@
 - reduced memory usage for urban density calculation, see #2828
 - urban density is now based on road junctions, so the according parameters need adjustment in case
   the config file does not use the defaults, see #2842
+- removed heading penalty *time*, see #2563
+- base graph no longer allows loop edges, see #2862
 
 ### 7.0 [14 Mar 2023]